+++ NEWS.md
@@ -1,5 +1,8 @@
 # testthat 0.11.0.9000
 
+* `throws_error` and `expect_error` can now take a `class` argument to express
+   an expected subclass for the error condition
+
 * Fix failure message for `throws_error` in case where no error is raised (@nealrichardson, #342).
 
 * Added [Catch](https://github.com/philsquared/Catch) for unit testing of C++ code.